Added spikes!
Yes, I know it's been a long time since the last commit. I was partially distracted by Minecraft (I know right) and the trickiness of the triangle hitbox of a spike. In the end, I came up with a compromise. (Note: Might have to redo the hitboxes so it's not so obvious that I approximated the triangle with rectangles.)